Holy cow!
A bunch of work that I've been pushing off to.... well push. Rewrote a bunch of css code, added in a bunch of images, and added in some other NEWER code. Eventually going to downsize, but I'm happy with how it is for now (Gonna test in a live environment to see how that turned out)